requirements.txt
numpy
pandas
faker
reportlab
pdfplumber
pydantic
pyyaml
